feat: 快捷键独立设置页,并加上右侧面板开关与新建/关闭 tab
