+++ src/main/services/capabilityGateway.ts
@@ -600,10 +600,13 @@ function parseAgentTypeFields(
 
 const DEFAULT_KEYBINDINGS: Record<string, string> = {
   'toggle-sidebar': 'mod+b',
+  'toggle-side-panel': 'mod+j',
   'open-settings': 'mod+,',
   'new-conversation': 'mod+n',
   'next-tab': process.platform === 'darwin' ? 'ctrl+tab' : 'mod+tab',
   'prev-tab': process.platform === 'darwin' ? 'ctrl+shift+tab' : 'mod+shift+tab',
+  'new-side-tab': 'mod+t',
+  'close-side-tab': 'mod+w',
 };
 
 export function createCapabilityHandlers(
